Add gogrs server
