--- cmd/tradingdays_server/docker-compose.yml
@@ -1,5 +1,0 @@
-tradingdaysserver:
-    image: toomore/gogrs-mini
-    command: tradingdays_server
-    ports:
-        - "80:59123"
